improved rendering performance
